Moved zone api commands to a separate file, and added some small doc changes
