+++ library/network/src/Makefile.am
@@ -38,13 +38,18 @@ yfwlibdir = @ylibdir@/y2firewall
 yfwlib_DATA = \
   lib/y2firewall/firewalld.rb
 
-yfwdlibdir = @ylibdir@/y2firewall/firewalld
+yfwdlibdir = @yfwdlibdir@/firewalld
 yfwdlib_DATA = \
   lib/y2firewall/firewalld/api.rb \
   lib/y2firewall/firewalld/relations.rb \
   lib/y2firewall/firewalld/zone.rb \
   lib/y2firewall/firewalld/zone_parser.rb
 
-EXTRA_DIST = $(module_DATA) $(scrconf_DATA) $(agent_SCRIPTS) $(ylib_DATA) $(yfwlib_DATA) $(yfwdlibdir_DATA)
+yfwdapilibdir = @yfwdlibdir@/api/
+yfwdapilib_DATA = \
+  lib/y2firewall/firewalld/api/services.rb \
+  lib/y2firewall/firewalld/api/zones.rb
+
+EXTRA_DIST = $(module_DATA) $(scrconf_DATA) $(agent_SCRIPTS) $(ylib_DATA) $(yfwlib_DATA) $(yfwdapilib_DATA)
 
 include $(top_srcdir)/Makefile.am.common
